[CLEANUP] Remove MultiWebDriver
It only creates problems. You can not separate the processes on the server when running multiple browsers, so all non-idempotent actions had to be wrapped in inFirstDriver or so.

Migration: Copy your test suite and use firefox on the seconds suite.
There might be a better way in the future, but not as of now.